Changed DomesticAnimal and NinjaMonkey
Added parameter player to moveToStealedPlayer so that it follows the
player to steal from.
Added line to getNearestPlayer so that nearest player can't be
yourself(owner).
Edited playersNearby so that the range is only 5 squares.
Implement override of logic that uses playersNearby and
movesToStealedPlayer to make the ninja monkey move sometimes only.